relocate ratings sample data
